# lib/api-zod/src/generated/types/agentQueryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface AgentQueryResponse {
  query: string;
  agent_name: string;
  agent_id: string;
  detected_domains: string[];
  response: string;
  conversation_history: Message[];
  skills_used?: string[];
  tools_used?: string[];
  processing_time_ms: number;
}