# lib/api-zod/src/generated/types/evalRunResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvalRunResultCategoryScores } from "./evalRunResultCategoryScores";
import type { EvalTestResult } from "./evalTestResult";

export interface EvalRunResult {
  run_id: string;
  total_tests: number;
  passed: number;
  failed: number;
  pass_rate: number;
  avg_latency_ms: number;
  category_scores: EvalRunResultCategoryScores;
  results: EvalTestResult[];
  ran_at: string;
}