# lib/api-zod/src/generated/types/evalTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvalTestResult {
  test_id: string;
  query: string;
  category: string;
  passed: boolean;
  score: number;
  expected_domain: string;
  actual_domain: string;
  expected_agent: string;
  actual_agent: string;
  response_preview: string;
  latency_ms: number;
  error?: string | null;
}