# lib/api-zod/src/generated/types/llmSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LlmSettingsEnvHints } from "./llmSettingsEnvHints";
import type { LlmSettingsProvider } from "./llmSettingsProvider";

export interface LlmSettings {
  /** Active LLM provider */
  provider: LlmSettingsProvider;
  /** Currently configured model name */
  model: string;
  /** Base URL for OpenAI-compatible providers */
  base_url?: string | null;
  /** Whether API key is configured */
  has_api_key: boolean;
  /** Whether Gemini API is configured */
  gemini_configured: boolean;
  /** Whether OpenAI-compatible provider is configured */
  openai_configured: boolean;
  /** List of available providers */
  available_providers: string[];
  /** Hints for configuring providers */
  env_hints: LlmSettingsEnvHints;
}