# lib/api-zod/src/generated/types/queryAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QueryAnalysis {
  query: string;
  recommended_agent: string;
  confidence: number;
  detected_domains?: string[];
}